+++ tests/test_dynamic_functions.py
@@ -16,6 +16,12 @@ def test_doc_dynamic_functions(app, status, warning):
     assert sum(1 for _ in re.finditer('<span class="needs-tag test">test</span>', html)) == 2
     assert sum(1 for _ in re.finditer('<span class="needs-tag my_tag">my_tag</span>', html)) == 1
 
+    assert sum(1 for _ in re.finditer('<span class="needs-tag test_4a">test_4a</span>', html)) == 1
+    assert sum(1 for _ in re.finditer('<span class="needs-tag test_4b">test_4b</span>', html)) == 1
+    assert sum(1 for _ in re.finditer('<span class="needs-tag TEST_4">TEST_4</span>', html)) == 1
+
+    assert sum(1 for _ in re.finditer('<span class="needs-tag TEST_5">TEST_5</span>', html)) == 1
+
     assert 'Test output of need TEST_3. args:' in html