+++ tests/test_global_options.py
@@ -14,7 +14,13 @@ def test_doc_global_option(app, status, warning):
     assert 'global_1' in html
     assert 'global_2' in html
     assert 'global_3' in html
+    assert 'global_4' in html
+    assert 'global_5' in html
 
     assert 'test_global' in html
     assert '1.27' in html
     assert 'Test output of need GLOBAL_ID' in html
+
+    assert 'STATUS_IMPL' in html
+    assert 'STATUS_UNKNOWN' in html
+    assert 'STATUS_CLOSED' not in html