+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 -----
 
 * Improvement: Role :ref:`role_need` supports standard sphinx-ref syntax. Example: ``:need:`custom name <need_id>```
+* Improvement: Added :ref:`global_option_filters` to set values of global options only under custom circumstances.
+* Bugfix: Fixed corrupted :ref:`dynamic_functions` handling for ``tags`` and other list options.
+  (`#100 <https://github.com/useblocks/sphinxcontrib-needs/issues/100>`_)
 
 0.4.2
 -----